Remove Kafka event handler as source of order workflow and clean up for conference presentation
