+++ typescript/food-ordering/app/src/delivery-app/external/driver_mobile_app_sim_utils.ts
