+++ typescript/food-ordering/docker-compose-base.yml
@@ -47,18 +47,10 @@ services:
     environment:
       - COLLECTOR_OTLP_ENABLED=true
 
-  order_app:
-    container_name: order_app
-    build:
-      context: ./app
-      dockerfile: Dockerfile-order
-    environment:
-      - RESTATE_DEBUG_LOGGING=INVOKE
-      - RESTAURANT_ENDPOINT=http://restaurant_pos:5050
-      - KAFKA_BOOTSTRAP_SERVERS=broker:29092
-
   delivery_app:
     container_name: delivery_app
+    ports:
+      - "9081:9081"
     build:
       context: ./app
       dockerfile: Dockerfile-delivery
@@ -72,48 +64,15 @@ services:
     build:
       context: ./app
       dockerfile: Dockerfile-restaurant
-    environment:
-      - RESTATE_RUNTIME_ENDPOINT=http://runtime:8080
-
-  runtime:
-    image: docker.io/restatedev/restate:0.7
-    depends_on:
-      - order_app
-      - delivery_app
-      - restaurant_pos
-      - broker
-      - jaeger
     ports:
-      - "9070:9070"
-      - "9071:9071"
-      - "8080:8080"
-    volumes:
-      - ./restate-docker.yaml:/restate.yaml
+      - "5050:5050"
     environment:
-      - RESTATE_OBSERVABILITY__TRACING__ENDPOINT=http://jaeger:4317
-      - RESTATE_CONFIG=/restate.yaml
-
-  runtimesetup:
-    image: alpine
-    depends_on:
-      - runtime
-      - order_app
-      - delivery_app
-    restart: "no"
-    entrypoint: ["sh", "-c", "sleep 5 && 
-    apk add --no-cache bash jq curl &&
-    curl -X POST 'runtime:9070/deployments' -H 'content-type: application/json' -d '{\"uri\": \"http://order_app:9080\"}' &&
-    curl -X POST 'runtime:9070/deployments' -H 'content-type: application/json' -d '{\"uri\": \"http://delivery_app:9081\"}' &&
-    sleep 3 && 
-    curl -X POST 'runtime:9070/subscriptions' -H 'content-type: application/json' -d '{ \"source\":\"kafka://my-cluster/driver-updates\", \"sink\":\"service://driver-digital-twin/handleDriverLocationUpdateEvent\" }' &&
-    curl -X POST -H 'content-type: application/json' runtime:8080/driver-mobile-app/startDriver -d '{\"key\": \"driver-01\", \"request\": {} }' &&
-    curl -X POST -H 'content-type: application/json' runtime:8080/driver-mobile-app/startDriver -d '{\"key\": \"driver-02\", \"request\": {} }' &&
-    exit 1"]
+      - RESTATE_RUNTIME_ENDPOINT=http://host.docker.internal:8080
+    extra_hosts:
+      - "host.docker.internal:host-gateway"
 
   webui:
     build:
       context: ./webui
-    depends_on:
-      - runtimesetup
     ports:
       - "3000:3000"
