+++ typescript/food-ordering/restate-local.yaml
@@ -1,0 +1,10 @@
+worker:
+  kafka:
+    clusters:
+      my-cluster:
+        bootstrap.servers: localhost:9092
+  storage_query:
+    bind_address: 0.0.0.0:5432
+  invoker:
+    retry_policy:
+      max_interval: 2s
